# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API v2",
        "contact": {
            "email": "ybier@esme-solutions.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sites": {
            "get": {
                "tags": [
                    "Sites"
                ],
                "summary": "Get a list of sites",
                "operationId": "b2375423d59eedc5dcce02e3cb15173a",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sites"
                ],
                "summary": "Create a site",
                "operationId": "4fe9547a6429a6b8f5b442d220939790",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ben Site"
                                    },
                                    "key": {
                                        "type": "string",
                                        "example": "010000BEN1"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Baker St"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/sites/{id}": {
            "get": {
                "tags": [
                    "Sites"
                ],
                "summary": "Get details of a site",
                "operationId": "5e74356f1cbbbdbe769db68482cc3a85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the site to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sites"
                ],
                "summary": "Update a site",
                "operationId": "a3231ef301930fefe403ab4a7b506612",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the site to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Ben Site"
                                            },
                                            "key": {
                                                "type": "string",
                                                "example": "010000BEN1"
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "123 Baker St"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sites"
                ],
                "summary": "Delete a site",
                "operationId": "d6cffc5b1111fb5ede331beb6ba025f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the site to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful deleted"
                    },
                    "401": {
                        "description": "Unauthorized Access"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "clientCredentials": {
                        "authorizationUrl": "http://powerbat-laravel/oauth/authorize",
                        "tokenUrl": "http://powerbat-laravel/oauth/token",
                        "refreshUrl": "http://powerbat-laravel/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}